0.0.5

+++ structs/economy.go
@@ -2,6 +2,6 @@ package structs
 
 type EconomyUser struct {
 	ID string `rethinkdb:"id"`
-	COIN int
-	BANK int
+	COIN int64
+	BANK int64
 }
